+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 - Technical - Upgrade to the `forest-express` v4 public dependency.
 - Technical - Apply eslint rules to `has-many-dissociator`.
+- Technical - Apply eslint rules to `resource-remover`.
 
 ## RELEASE 4.0.0 - 2019-10-04
 ### Added